requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.2
MarkupSafe==2.1.3
multidict==6.0.4
requests==2.28.2
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
